# packages/server/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
...
			"IStandardResponse_any_": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMultiStreamDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"streamId": {
						"type": "string"
					},
					"targetStreamKey": {
						"type": "string"
					},
					"targetURL": {
						"type": "string"
					},
					"organizationId": {
						"type": "string"
					},
					"socialId": {
						"type": "string"
					},
					"socialType": {
						"type": "string"
					},
					"broadcastId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"streamId",
					"targetStreamKey",
					"targetURL",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse_void_": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteMultiStreamDto": {
				"properties": {
					"streamId": {
						"type": "string"
					},
					"targetId": {
						"type": "string"
					},
					"organizationId": {
						"type": "string"
					}
				},
				"required": [
					"streamId",
					"targetId",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse__url-string--assetId-string__": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"assetId": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"assetId",
							"url"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse_string_": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse__playbackUrl-string--phaseStatus-string__": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"phaseStatus": {
								"type": "string"
							},
							"playbackUrl": {
								"type": "string"
							}
						},
						"required": [
							"phaseStatus",
							"playbackUrl"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse__viewCount-number--playTimeMins-number__": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"playTimeMins": {
								"type": "number",
								"format": "double"
							},
							"viewCount": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"playTimeMins",
							"viewCount"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateClipDto": {
				"properties": {
					"playbackId": {
						"type": "string"
					},
					"sessionId": {
						"type": "string"
					},
					"recordingId": {
						"type": "string"
					},
					"start": {
						"type": "number",
						"format": "double"
					},
					"end": {
						"type": "number",
						"format": "double"
					},
					"organizationId": {
						"type": "string"
					},
					"stageId": {
						"type": "string"
					},
					"isEditorEnabled": {
						"type": "boolean"
					},
					"editorOptions": {
						"properties": {
							"captionColor": {
								"type": "string"
							},
							"captionFont": {
								"type": "string"
							},
							"captionLinesPerPage": {
								"type": "number",
								"format": "double"
							},
							"captionPosition": {
								"type": "string"
							},
							"captionEnabled": {
								"type": "boolean"
							},
							"selectedAspectRatio": {
								"type": "string"
							},
							"events": {
								"items": {
									"properties": {
										"sessionId": {
											"type": "string"
										},
										"label": {
											"type": "string"
										}
									},
									"required": [
										"sessionId",
										"label"
									],
									"type": "object"
								},
								"type": "array"
							},
							"frameRate": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"captionColor",
							"captionFont",
							"captionLinesPerPage",
							"captionPosition",
							"captionEnabled",
							"selectedAspectRatio",
							"events",
							"frameRate"
						],
						"type": "object"
					},
					"clipSessionId": {
						"type": "string"
					},
					"clipEditorId": {
						"type": "string"
					}
				},
				"required": [
					"playbackId",
					"sessionId",
					"recordingId",
					"start",
					"end",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStandardResponse__type-string--url-string__": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"url": {
								"type": "string"
							},
							"type": {
								"type": "string"
							}
						},
						"required": [
							"url",
							"type"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"StageType": {
				"enum": [
					"custom",
					"livepeer"
				],
				"type": "string"
			},
...
			"IStandardResponse_IStage_": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IStage"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStageDto": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"eventId": {
						"type": "string"
					},
					"streamSettings": {
						"$ref": "#/components/schemas/IStreamSettings"
					},
					"plugins": {
						"items": {
							"$ref": "#/components/schemas/IPlugin"
						},
						"type": "array"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"slug": {
						"type": "string"
					},
					"published": {
						"type": "boolean"
					},
					"isMultipleDate": {
						"type": "boolean"
					},
					"organizationId": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"streamDate": {
						"type": "string",
						"format": "date-time"
					},
					"streamEndDate": {
						"type": "string",
						"format": "date-time"
					},
					"mintable": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"nftCollections": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						]
					},
					"recordingIndex": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"$ref": "#/components/schemas/StageType"
					},
					"transcripts": {
						"properties": {
							"text": {
								"type": "string"
							},
							"chunks": {
								"items": {
									"properties": {
										"word": {
											"type": "string"
										},
										"end": {
											"type": "number",
											"format": "double"
										},
										"start": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"word",
										"end",
										"start"
									],
									"type": "object"
								},
								"type": "array"
							},
							"subtitleUrl": {
								"type": "string"
							},
							"lastSegmentTimestamp": {
								"type": "number",
								"format": "double"
							},
							"status": {
								"$ref": "#/components/schemas/TranscriptionStatus"
							}
						},
						"required": [
							"text",
							"chunks",
							"subtitleUrl",
							"lastSegmentTimestamp",
							"status"
						],
						"type": "object"
					},
					"source": {
						"properties": {
							"type": {
								"type": "string"
							},
							"m3u8Url": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"m3u8Url",
							"url"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"organizationId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateHlsStageDto": {
				"properties": {
					"_id": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"eventId": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							},
							{
								"type": "string"
							}
						]
					},
					"streamSettings": {
						"$ref": "#/components/schemas/IStreamSettings"
					},
					"plugins": {
						"items": {
							"$ref": "#/components/schemas/IPlugin"
						},
						"type": "array"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"slug": {
						"type": "string"
					},
					"published": {
						"type": "boolean"
					},
					"isMultipleDate": {
						"type": "boolean"
					},
					"organizationId": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"streamDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"streamEndDate": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "string"
							}
						]
					},
					"mintable": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					},
					"nftCollections": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/mongoose.Types.ObjectId"
							},
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						]
					},
					"recordingIndex": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"$ref": "#/components/schemas/StageType"
					},
					"transcripts": {
						"properties": {
							"text": {
								"type": "string"
							},
							"chunks": {
								"items": {
									"properties": {
										"word": {
											"type": "string"
										},
										"end": {
											"type": "number",
											"format": "double"
										},
										"start": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"word",
										"end",
										"start"
									],
									"type": "object"
								},
								"type": "array"
							},
							"subtitleUrl": {
								"type": "string"
							},
							"lastSegmentTimestamp": {
								"type": "number",
								"format": "double"
							},
							"status": {
								"$ref": "#/components/schemas/TranscriptionStatus"
							}
						},
						"required": [
							"text",
							"chunks",
							"subtitleUrl",
							"lastSegmentTimestamp",
							"status"
						],
						"type": "object"
					},
					"source": {
						"properties": {
							"type": {
								"type": "string"
							},
							"m3u8Url": {
								"type": "string"
							},
							"url": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"m3u8Url",
							"url"
						],
						"type": "object"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"organizationId",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IStandardResponse_Array_IStage__": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IStage"
						},
						"type": "array"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"name": "Authorization",
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "streameth-api",
		"version": "3.0.0",
		"contact": {}
	},
	"paths": {
...
		"/streams/multistream": {
			"post": {
				"operationId": "CreateMultiStream",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Create Multistream",
				"tags": [
					"Stream"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMultiStreamDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMultiStream",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_void_"
								}
							}
						}
					}
				},
				"summary": "Delete Multistream",
				"tags": [
					"Stream"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteMultiStreamDto"
							}
						}
					}
				}
			}
		},
		"/streams/asset": {
			"post": {
				"operationId": "CreateAsset",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse__url-string--assetId-string__"
								}
							}
						}
					}
				},
				"summary": "Create stream asset",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/streams/{streamId}": {
			"get": {
				"operationId": "GetStream",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Get Stream",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "streamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/asset/{assetId}": {
			"get": {
				"operationId": "GetAsset",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Get Asset",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/asset/url/{assetId}": {
			"get": {
				"operationId": "GetVideoUrl",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_string_"
								}
							}
						}
					}
				},
				"summary": "Get Video url",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/asset/phase-action/{assetId}": {
			"get": {
				"operationId": "GetPhaseAction",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse__playbackUrl-string--phaseStatus-string__"
								}
							}
						}
					}
				},
				"summary": "Get Video Phase Action",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/metric/{playbackId}": {
			"get": {
				"operationId": "GetSessionMetrics",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse__viewCount-number--playTimeMins-number__"
								}
							}
						}
					}
				},
				"summary": "Get stream metrics",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "playbackId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/recording/{streamId}": {
			"get": {
				"operationId": "GetStreamRecordings",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Get stream recordings",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "streamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/upload/{assetId}": {
			"get": {
				"operationId": "UploadToIpfs",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_string_"
								}
							}
						}
					}
				},
				"summary": "Upload",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/streams/clip": {
			"post": {
				"operationId": "CreateClip",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Create clip",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateClipDto"
							}
						}
					}
				}
			}
		},
		"/streams/thumbnail/generate": {
			"post": {
				"operationId": "GenerateThumbnail",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_any_"
								}
							}
						}
					}
				},
				"summary": "Generate thumbnail",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/streams/hls": {
			"post": {
				"operationId": "GetHls",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse__type-string--url-string__"
								}
							}
						}
					}
				},
				"summary": "Get HLS url",
				"tags": [
					"Stream"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									}
								},
								"required": [
									"url"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/stages": {
			"post": {
				"operationId": "CreateStage",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_IStage_"
								}
							}
						}
					}
				},
				"summary": "Create Stage",
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStageDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllStages",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_Array_IStage__"
								}
							}
						}
					}
				},
				"summary": "Get All Stage",
				"tags": [
					"Stage"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "published",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/stages/hls": {
			"post": {
				"operationId": "CreateHlsStage",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_IStage_"
								}
							}
						}
					}
				},
				"summary": "Create HLS Stage",
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateHlsStageDto"
							}
						}
					}
				}
			}
		},
		"/stages/{stageId}": {
			"put": {
				"operationId": "EditStage",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_IStage_"
								}
							}
						}
					}
				},
				"summary": "Update Stage",
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "stageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateStageDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetStageById",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_IStage_"
								}
							}
						}
					}
				},
				"summary": "Get Stage by id",
				"tags": [
					"Stage"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "stageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteStage",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_void_"
								}
							}
						}
					}
				},
				"summary": "Delete Stage",
				"tags": [
					"Stage"
				],
				"security": [
					{
						"jwt": [
							"org"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "stageId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OrgIdDto"
							}
						}
					}
				}
			}
		},
		"/stages/event/{eventId}": {
			"get": {
				"operationId": "GetAllStagesForEvent",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_Array_IStage__"
								}
							}
						}
					}
				},
				"summary": "Get All Stage for Event",
				"tags": [
					"Stage"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "eventId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/stages/organization/{organizationId}": {
			"get": {
				"operationId": "GetAllStagesForOrganization",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IStandardResponse_Array_IStage__"
								}
							}
						}
					}
				},
				"summary": "Get All Stage for Organization",
				"tags": [
					"Stage"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "organizationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "fromDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "untilDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/streameth-platform-packages-new2"
		}
	]
}